move files to separate folder for safekeeping
